backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "motor>=3.7.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
indent-width = 2
